@@ -1,20 +1,21 @@
 <Project Sdk="Microsoft.NET.Sdk.Web">
-  <PropertyGroup>
-    <TargetFramework>netcoreapp2.0</TargetFramework>
-    <DocumentationFile>bin\Debug\netcoreapp2.0\Sample.AspNetCoreWebApi.xml</DocumentationFile>
-    <NoWarn>1701;1702;1705;1591</NoWarn>
-  </PropertyGroup>
-  <ItemGroup>
-    <Folder Include="wwwroot\" />
-  </ItemGroup>
-  <ItemGroup>
-    <PackageReference Include="Mapster" Version="3.1.8" />
-    <PackageReference Include="Microsoft.AspNetCore.All" Version="2.0.5" />
-    <PackageReference Include="Kros.KORM" Version="3.6.3" />
-    <PackageReference Include="Kros.Utils" Version="1.5.1" />
-    <PackageReference Include="Swashbuckle.AspNetCore" Version="2.2.0" />
-  </ItemGroup>
-  <ItemGroup>
-    <DotNetCliToolReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Tools" Version="2.0.2" />
-  </ItemGroup>
+	<PropertyGroup>
+		<TargetFramework>net9.0</TargetFramework>
+		<NoWarn>1701;1702;1705;1591</NoWarn>
+		<DocumentationFile>bin\Debug\netcoreapp2.0\Sample.AspNetCoreWebApi.xml</DocumentationFile>
+	</PropertyGroup>
+	<ItemGroup>
+		<Folder Include="wwwroot\" />
+	</ItemGroup>
+	<ItemGroup>
+		<PackageReference Include="Kros.KORM" Version="7.3.0" />
+		<PackageReference Include="Kros.KORM.Extensions.Asp" Version="2.0.1" />
+		<PackageReference Include="Kros.Utils" Version="3.1.0" />
+		<PackageReference Include="Mapster" Version="7.4.0" />
+		<PackageReference Include="Microsoft.AspNetCore.Authentication.JwtBearer" Version="9.0.9" />
+		<PackageReference Include="Swashbuckle.AspNetCore" Version="9.0.6" />
+	</ItemGroup>
+	<ItemGroup>
+		<DotNetCliToolReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Tools" Version="2.0.2" />
+	</ItemGroup>
 </Project>
