@@ -1,8 +1,8 @@
-using System;
-using System.IO;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.Extensions.DependencyInjection;
-using Swashbuckle.AspNetCore.Swagger;
+using Microsoft.OpenApi.Models;
+using System;
+using System.IO;
 
 namespace Sample.AspNetCoreWebApi.Services
 {
@@ -18,27 +18,27 @@ namespace Sample.AspNetCoreWebApi.Services
         public static IServiceCollection AddSwaggerDocumentation(this IServiceCollection services) =>
             services.AddSwaggerGen(c =>
             {
-                c.SwaggerDoc("v1", new Info
+                c.SwaggerDoc("v1", new OpenApiInfo
                 {
                     Version = "v1",
-                        Title = "People API",
-                        Description = "A simple example ASP.NET Core Web API",
-                        Contact = new Contact
-                        {
-                            Name = "Miňo Martiniak",
-                                Url = "https://twitter.com/MinoMartiniak"
-                        }
+                    Title = "People API",
+                    Description = "A simple example ASP.NET Core Web API",
+                    Contact = new OpenApiContact
+                    {
+                        Name = "Miňo Martiniak",
+                        Url = new Uri("https://twitter.com/MinoMartiniak")
+                    }
                 });
                 var basePath = AppContext.BaseDirectory;
                 var xmlPath = Path.Combine(basePath, "Sample.AspNetCoreWebApi.xml");
                 c.IncludeXmlComments(xmlPath);
 
-                c.AddSecurityDefinition("Bearer", new ApiKeyScheme
+                c.AddSecurityDefinition("Bearer", new OpenApiSecurityScheme
                 {
                     Description = "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\"",
-                        Name = "Authorization",
-                        In = "header",
-                        Type = "apiKey"
+                    Name = "Authorization",
+                    In = ParameterLocation.Header,
+                    Type = SecuritySchemeType.ApiKey
                 });
             });
 
